requirements.txt
fastapi
uvicorn[standard]
python-multipart

pydantic[email]
pydantic-settings

sqlalchemy
psycopg2-binary
alembic

celery
redis

jinja2

beautifulsoup4
bleach

geoip2
user-agents
aiosmtplib

python-dotenv
loguru

pytest
httpx
openpyxl

supabase

passlib[bcrypt]
bcrypt<4.1
python-jose[cryptography]

